fix(appliance): do not delete unowned resources
The appliance checks owner references, and only deletes (or updates)
resources whose owner references matches the configmap being reconciled.

The current user interface to external databases, is for the admin to
create secrets with a well-known name out of band (e.g. "pgsql-auth")
and then disable the relevant backing services (e.g. pgsql). This commit
fixes a bug in which the appliance would have deleted such secrets,
leaving the admin unable to use external databases.

Discovered while investigating
https://linear.app/sourcegraph/issue/REL-14/ensure-pods-roll-when-referenced-secret-changes,
but does not close it.

+++ internal/appliance/reconciler/kubernetes.go
@@ -28,7 +28,7 @@ func reconcileObject[T client.Object](
 	sg *config.Sourcegraph, owner client.Object,
 ) error {
 	if cfg.IsDisabled() {
-		return r.ensureObjectDeleted(ctx, obj)
+		return ensureObjectDeleted(ctx, r, owner, obj)
 	}
 
 	updateIfChanged := struct {
@@ -102,6 +102,11 @@ func createOrUpdateObject[R client.Object](
 		return err
 	}
 
+	if !isControlledBy(owner, existingRes) {
+		logger.Info("refusing to update non-owned resource")
+		return nil
+	}
+
 	if cfgHash != existingRes.GetAnnotations()[config.AnnotationKeyConfigHash] {
 		logger.Info("Found existing object with spec that does not match desired state. Clobbering it.")
 		if err := r.Client.Update(ctx, obj); err != nil {
@@ -125,10 +130,28 @@ func isNamespaced(obj client.Object) bool {
 	return false
 }
 
-func (r *Reconciler) ensureObjectDeleted(ctx context.Context, obj client.Object) error {
+func ensureObjectDeleted[T client.Object](ctx context.Context, r *Reconciler, owner client.Object, obj T) error {
+	// We need to try to get the object first, in order to check its owner
+	// references later.
+	objKey := types.NamespacedName{Name: obj.GetName(), Namespace: obj.GetNamespace()}
+	if err := r.Client.Get(ctx, objKey, obj); err != nil {
+		if kerrors.IsNotFound(err) {
+			// Object doesn't exist, we don't need to delete it
+			return nil
+		}
+	}
+
 	logger := log.FromContext(ctx).WithValues("kind", obj.GetObjectKind().GroupVersionKind(), "namespace", obj.GetNamespace(), "name", obj.GetName())
+
+	if !isControlledBy(owner, obj) {
+		logger.Info("refusing to delete non-owned resource")
+		return nil
+	}
+
+	logger.Info("deleting resource")
 	if err := r.Client.Delete(ctx, obj); err != nil {
 		if kerrors.IsNotFound(err) {
+			// If by chance it got deleted concurrently, no harm done.
 			return nil
 		}
 
@@ -136,6 +159,15 @@ func (r *Reconciler) ensureObjectDeleted(ctx context.Context, obj client.Object)
 		return err
 	}
 	return nil
+}
+
+func isControlledBy(owner, obj client.Object) bool {
+	for _, ownerRef := range obj.GetOwnerReferences() {
+		if owner.GetUID() == ownerRef.UID {
+			return true
+		}
+	}
+	return false
 }
 
 func configHash(configElement any) (string, error) {
